+++ src/Viserio/Component/Parsers/Parsers/PhpParser.php
@@ -1,12 +1,12 @@
 <?php
 declare(strict_types=1);
-namespace Viserio\Component\Parsers\Formats;
+namespace Viserio\Component\Parsers\Parsers;
 
 use Throwable;
 use Viserio\Component\Contracts\Parsers\Exception\ParseException;
-use Viserio\Component\Contracts\Parsers\Format as FormatContract;
+use Viserio\Component\Contracts\Parsers\Parser as ParserContract;
 
-class PhpParser implements FormatContract
+class PhpParser implements ParserContract
 {
     /**
      * {@inheritdoc}
