+++ src/Viserio/Component/Parsers/Parsers/QueryStrParser.php
@@ -1,10 +1,10 @@
 <?php
 declare(strict_types=1);
-namespace Viserio\Component\Parsers\Formats;
+namespace Viserio\Component\Parsers\Parsers;
 
-use Viserio\Component\Contracts\Parsers\Format as FormatContract;
+use Viserio\Component\Contracts\Parsers\Parser as ParserContract;
 
-class QueryStrParser implements FormatContract
+class QueryStrParser implements ParserContract
 {
     /**
      * {@inheritdoc}
