+++ src/Viserio/Component/Parsers/Parsers/XmlParser.php
@@ -1,13 +1,13 @@
 <?php
 declare(strict_types=1);
-namespace Viserio\Component\Parsers\Formats;
+namespace Viserio\Component\Parsers\Parsers;
 
 use InvalidArgumentException;
 use Viserio\Component\Contracts\Parsers\Exception\ParseException;
-use Viserio\Component\Contracts\Parsers\Format as FormatContract;
+use Viserio\Component\Contracts\Parsers\Parser as ParserContract;
 use Viserio\Component\Parsers\Utils\XmlUtils;
 
-class XmlParser implements FormatContract
+class XmlParser implements ParserContract
 {
     /**
      * {@inheritdoc}
